Fixed #100 Add loot table and update interactions for HopsCropBlock
Created a new loot table for Hops Vines that uses age property to determine the loot. Also updated the HopsCropBlock to decouple it from the Config and to use the loot table for determining drops. This provides more flexibility for changing the behavior of hop vine drops in the future.