Add light emission feature to BrewKettleBlock
A lightLevel variable has been introduced in the BrewKettleBlock class in order to support the feature of light emission. The light emission of the BrewKettleBlock will now be determined by the block's lit state. This provides visual feedback that the kettle is processing by emitting a custom light level when lit. The light level value is fetched from the GrowthcraftCellarConfig.