#100 Mark config classes as deprecated and add comments in GrowthcraftMilkConfig
The config classes of bamboo, apples and rice have been marked as deprecated. Detailed comments have been added to the GrowthcraftMilkConfig functions, clarifying the intention and usage of each method. A TODO item has also been added to the GrowthcraftConfig class for the implementation of a feature allowing server admins to prevent salt generation.