Progress #124 Add foil property to GrowthCraft items
This update adds a new boolean property 'hasFoil' to the GrowthcraftItem class in Growthcraft, which signifies whether an item has foil or not—the foil attribute is used to differentiate certain items. The hasFoil property is used in constructing methods and property initialisation. Some items in the GrowthcraftCellarItems have been updated to have the foil property.